refactor: extract GameList hooks
Closes #97
